requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.45.0
datasets>=2.18.0
accelerate>=0.27.0
peft>=0.10.0
bitsandbytes>=0.43.0
scipy>=1.12.0
scikit-learn>=1.4.0
av>=12.0.0
pillow>=10.0.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=4.25.0
wandb>=0.16.0
huggingface-hub>=0.21.0
safetensors>=0.4.0
tokenizers>=0.15.0
einops>=0.7.0
